+++ EsKitsLibrary/src/main/java/io/github/xuyao5/dkl/eskits/context/translator/TwoArgEventTranslator.java
@@ -1,4 +1,4 @@
-package io.github.xuyao5.dkl.eskits.context.event;
+package io.github.xuyao5.dkl.eskits.context.translator;
 
 import com.lmax.disruptor.EventTranslatorTwoArg;
 import lombok.NonNull;
@@ -8,7 +8,7 @@ import lombok.NonNull;
  * @version 5/07/21 00:52
  */
 @FunctionalInterface
-public interface TwoArgEvent<E> {
+public interface TwoArgEventTranslator<E> {
 
     <A, B> void translate(@NonNull EventTranslatorTwoArg<E, A, B> translator, @NonNull A arg0, @NonNull B arg1);
 }
